+++ package.json
@@ -10,6 +10,7 @@
     "@types/node": "^16.11.35",
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.4",
+    "@types/react-router-dom": "^5.3.3",
     "axios": "^0.27.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
