+++ src/components/Sidebar.tsx
@@ -1,0 +1,100 @@
+import React, {useEffect, useState} from "react";
+import styled from 'styled-components'
+import SidebarItems from "./SidebarItems";
+import {Link} from "react-router-dom";
+
+// @ts-ignore
+function Sidebar(props, {defaultActive,}) {
+  const location = props.history.location;
+  const lastActiveIndexString = localStorage.getItem("lastActiveIndex");
+  const lastActiveIndex = Number(lastActiveIndexString);
+  const [activeIndex, setActiveIndex] = useState(lastActiveIndex || defaultActive);
+
+  // @ts-ignore
+  function changeActiveIndex(newIndex) {
+    localStorage.setItem("lastActiveIndex", newIndex)
+    setActiveIndex(newIndex)
+  }
+
+  // @ts-ignore
+  function getPath(path) {
+    if (path.charAt(0) !== "/") {
+      return  "/" + path;
+    }
+    return path;
+  }
+
+  useEffect(()=> {
+    const activeItem = SidebarItems.findIndex(item=> getPath(item.route) === getPath(location.pathname))
+    changeActiveIndex(activeItem);
+  }, [location])
+
+  return (
+    <>
+      <SidebarParent>
+        <div style={{position: 'fixed'}}>
+          {
+            SidebarItems.map((item, index)=> {
+              return (
+                <Link to={item.route}>
+                  <SidebarItem key={item.name}
+                    // @ts-ignore
+                    active={index === activeIndex}
+                  >
+                    <p>{item.name}</p>
+                  </SidebarItem>
+                </Link>
+              );
+            })
+          }
+        </div>
+        <div className="behind-the-scenes"/>
+      </SidebarParent>
+    </>
+  );
+}
+
+export default Sidebar;
+
+const SidebarParent = styled.div`
+  background: #cf3d2a;
+  
+  a {
+    text-decoration: none;
+  }
+  
+  & > div {
+    width: 250px;
+    height: 100vh;
+  }
+  
+  .behind-the-scenes {
+    width: 250px;
+    
+  }
+`;
+
+const SidebarItem = styled.div`
+  padding: 16px 24px;
+  transition: all 0.25s ease-in-out;
+  background: ${
+    // @ts-ignore
+    props => props.active ? "#b15b00" : ""
+  };
+  margin: 4px 12px;
+  border-radius: 4px;
+
+  p {
+    color: white;
+    font-weight: bold;
+    text-decoration: none;
+  }
+  
+  &:hover {
+    cursor:pointer;
+  }
+  
+  &:hover:not(:first-child) {
+    background: #c34a36;
+  }
+`;
